Split TrxResultStreamingStore into focused partials
The file had grown to 626 lines mixing five concerns. Split it into
partial-class files, each under 300 lines, with no API change:

- TrxResultStreamingStore.cs: public surface + lifecycle orchestration
- .BackgroundWriter.cs: queue, writer task, drain loop
- .InlineWriter.cs: synchronous WebAssembly path
- .FileIO.cs: batched writes, retries, truncation, file provisioning
- .Logging.cs: best-effort log helpers

The CA1416 suppression is now scoped to BackgroundWriter, the only file
that touches BlockingCollection<T>. IsInline is derived from the inline
buffer instead of the queue so the other partials never reach for a
browser-unsupported member.

Fixes #10332

Copilot-Session: d1b8caf3-4afa-4b42-a748-3a292d421d9e
